[feat] Update routes and functions to implement authentication

+++ server/model/db_file.js
@@ -1,35 +1,36 @@
 import { pool } from "./connection.js";
 // ----------------------------------------------------------
-const getDirId = async (parentId, dirName) => {
+const getDirId = async (userId, parentId, dirName) => {
   // console.log(`getDirId: parentId=${parentId} dirName=${dirName}`);
   const [row] = await pool.query(`
-    SELECT id FROM file WHERE parent_id = ? AND name = ? AND type = 'folder'
-  `, [parentId, dirName]);
+    SELECT id FROM file WHERE user_id = ? AND parent_id = ? AND name = ? AND type = 'folder'
+  `, [userId, parentId, dirName]);
   return row;
 };
 
-const saveMetadata = async(parentId, name, type, size) => {
+const saveMetadata = async(userId, parentId, name, type, size) => {
   const [row] = await pool.query(`
-    INSERT INTO file (parent_id, name, type, size) VALUES (?, ?, ?, ?) 
-  `, [parentId, name, type, size]);
+    INSERT INTO file (parent_id, name, type, size, user_id) VALUES (?, ?, ?, ?, ?) 
+  `, [parentId, name, type, size, userId]);
   return row;
 };
 
-const updMetadata = async(parentId, name, type, size) => {
+const updMetadata = async(userId, parentId, name, type, size) => {
   const [row] = await pool.query(`
-    UPDATE file SET size = ? WHERE parent_id = ? AND name = ? AND type = 'file'
-  `, [size, parentId, name, type]);
+    UPDATE file SET size = ? 
+    WHERE user_id = ? AND parent_id = ? AND name = ? AND type = 'file'
+  `, [userId, size, parentId, name, type]);
   return row;
 };
 
-const getFileList = async(parentId) => {
+const getFileList = async(userId, parentId) => {
   const [row] = await pool.query(`
-    SELECT id, name, type, size, updated_at FROM file WHERE parent_id = ?
-  `, parentId);
+    SELECT id, name, type, size, updated_at FROM file WHERE user_id = ? AND parent_id = ?
+  `, [userId, parentId]);
   return row;
 };
 
-const deleteWholeFolder = async(parentId) => {
+const deleteWholeFolder = async(userId, parentId) => {
   const connection = await pool.getConnection();
  
   try { 
@@ -39,13 +40,13 @@ const deleteWholeFolder = async(parentId) => {
 
     // Delete files
     await connection.query(`
-      DELETE FROM file WHERE parent_id = ?
-    `, parentId);
+      DELETE FROM file WHERE user_id = ? AND parent_id = ?
+    `, [userId, parentId]);
 
     // Delete folder
     await connection.query(`
-      DELETE FROM file WHERE id = ?
-    `, parentId);
+      DELETE FROM file WHERE user_id = ? AND id = ?
+    `, [userId, parentId]);
 
     // Complete transaction - Commit
     await connection.commit();
@@ -65,17 +66,17 @@ const deleteWholeFolder = async(parentId) => {
   }
 };
 
-const getFileId = async(parentId, fileName) => {
+const getFileId = async(userId, parentId, fileName) => {
   const [row] = await pool.query(`
-    SELECT id FROM file WHERE parent_id = ? AND name = ? AND type = 'file'
-  `, [parentId, fileName]);
+    SELECT id FROM file WHERE user_id = ? AND parent_id = ? AND name = ? AND type = 'file'
+  `, [userId, parentId, fileName]);
   return row;
 };
 
-const deleteById = async(id) => {
+const deleteById = async(userId, id) => {
   const [row] = await pool.query(`
-    DELETE FROM file WHERE id = ?
-  `, id);
+    DELETE FROM file WHERE user_id = ? AND id = ?
+  `, [userId, id]);
   console.log("file deleted");
   return row;
 };
